% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_learnr.R
\name{ci_learnr}
\alias{ci_learnr}
\title{Title}
\usage{
ci_learnr(name)
}
\arguments{
\item{name}{A string value containing the name of the tutorial you want to create.}
}
\value{
Creates a learnr tutorial with important components to create CI learnr tutorial.
}
\description{
A wrapper function that will create a learnr interactive document with the CSUCI theme.
}
